requirements.txt
pyyaml
jieba
snownlp
fredapi
pytrends
requests
beautifulsoup4
lxml
numpy
pandas
scipy
streamlit
plotly
